Code_Results/requirements.txt
openai>=1.0.0
pandas>=1.5.0
python-dotenv>=1.0.0
